% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{bearer_token}
\alias{bearer_token}
\title{Create the authorization header}
\usage{
bearer_token(account_set = load_account_set())
}
\arguments{
\item{account_set}{account set}
}
\value{
an httr header object with the authorization token
}
\description{
Create the authorization header
}
