% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{load_account_set}
\alias{load_account_set}
\title{Load an account set}
\usage{
load_account_set(account_set_name = "questrader", path = qt_dir())
}
\arguments{
\item{account_set_name}{the name of the stored account_set}

\item{path}{where to look for the account_set}
}
\value{
the account_set object
}
\description{
Load an account set
}
