% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_set.R
\name{new_account_set}
\alias{new_account_set}
\title{Create an account set}
\usage{
new_account_set(account_set_name, practice)
}
\arguments{
\item{account_set_name}{the name of the account set}

\item{practice}{boolean indicating whether the account is a practice account}
}
\value{
an object of class account_set
}
\description{
Create an account set
}
