% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{qt_add_account_set}
\alias{qt_add_account_set}
\title{Interactively add an account set}
\usage{
qt_add_account_set(directory = qt_dir())
}
\arguments{
\item{directory}{where to store the directory}
}
\value{
the created account set
}
\description{
Interactively add an account set
}
