% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qt_api_account_positions}
\alias{qt_api_account_positions}
\title{Get positions}
\usage{
qt_api_account_positions(account_id, account_set = load_account_set())
}
\arguments{
\item{account_id}{account ID.}

\item{account_set}{account set}
}
\value{
a list of tibbles with positions
}
\description{
Get positions
}
