% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qt_api_accounts}
\alias{qt_api_accounts}
\title{Get list of accounts}
\usage{
qt_api_accounts(account_set = load_account_set())
}
\arguments{
\item{account_set}{account set}
}
\value{
a tibble with a list of accounts
}
\description{
Gets the list of accounts corresponding to account_set
}
