% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qt_api_candles}
\alias{qt_api_candles}
\title{Get candlesticks}
\usage{
qt_api_candles(
  symbol_id,
  start_time,
  end_time,
  interval = c("OneMinute", "TwoMinutes", "ThreeMinutes", "FourMinutes", "FiveMinutes",
    "TenMinutese", "FifteenMinutes.", "TwentyMinutes", "HalfHour", "OneHour", "TwoHours",
    "FourHours", "OneDay", "OneWeek", "OneMonth", "OneYear"),
  account_set = load_account_set()
)
}
\arguments{
\item{symbol_id}{the ID of the symbol to look up.}

\item{start_time}{a datetime}

\item{end_time}{an end time}

\item{interval}{the time interval}

\item{account_set}{account set}
}
\value{
the current time
}
\description{
Get candlesticks
}
