% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qt_api_get}
\alias{qt_api_get}
\title{Get an API response}
\usage{
qt_api_get(
  name,
  account_set = load_account_set(),
  url_values = NULL,
  query = NULL,
  try = 0,
  auth_token = NULL
)
}
\arguments{
\item{name}{method name}

\item{account_set}{account set}

\item{url_values}{url values to glue into the url template, usually an ID}

\item{query}{a list to pass as the query}

\item{try}{how many times has this been called}

\item{auth_token}{a string to use for the access token. Usually this is not needed
but is used for testing.}
}
\value{
an httr response
}
\description{
Get an API response
}
