% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qt_api_search_symbol}
\alias{qt_api_search_symbol}
\title{Search symbol}
\usage{
qt_api_search_symbol(prefix, account_set = load_account_set())
}
\arguments{
\item{prefix}{prefix to search for}

\item{account_set}{account set}
}
\value{
a dataframe with symbols
}
\description{
Search symbol
}
