% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qt_api_symbols}
\alias{qt_api_symbols}
\title{Get information on symbols}
\usage{
qt_api_symbols(
  symbol_names = NULL,
  symbol_ids = NULL,
  account_set = load_account_set()
)
}
\arguments{
\item{symbol_names}{vector of exact names of symbols}

\item{symbol_ids}{vector of symbol ids}

\item{account_set}{account set}
}
\value{
a dataframe of symbol data.
}
\description{
Get information on symbols
}
\note{
Exactly one of symbol_names or symbol_ids should be provided. Not all data is provided by
this function—use .qt_api_symbols_raw to get all the data.
}
