% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{qt_api_time}
\alias{qt_api_time}
\title{Get the current time}
\usage{
qt_api_time(account_set = load_account_set(), auth_token = NULL)
}
\arguments{
\item{account_set}{account set}

\item{auth_token}{override access token. Usually not needed.}
}
\value{
the current time
}
\description{
Get the current time
}
