% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_set.R
\name{qt_default_account_set}
\alias{qt_default_account_set}
\title{Default account set filepath}
\usage{
qt_default_account_set()
}
\value{
The default account set directory
}
\description{
Default account set filepath
}
