% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{qt_init}
\alias{qt_init}
\title{Create the account_set directory}
\usage{
qt_init(directory = qt_dir())
}
\arguments{
\item{directory}{where to store account_set information}
}
\value{
The path to the created object (invisibly).
}
\description{
Create the account_set directory
}
