% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{qt_refresh_token}
\alias{qt_refresh_token}
\title{Refresh tokens}
\usage{
qt_refresh_token(
  account_set = load_account_set(),
  refresh_token = NULL,
  update_saved_account_set = TRUE
)
}
\arguments{
\item{account_set}{account_set object}

\item{refresh_token}{the value of a refresh token. Leave NULL to get this from keyring.}

\item{update_saved_account_set}{should the updated account set be updated? Almost always should be true.}
}
\value{
an updated account_set object
}
\description{
Submits a refresh token to Questrade to exchange for an access token, which can be used for
performing account operations. These tokens are stored securely in your platform's credential
store. This will also update the account_set object with the api server.
}
