% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{qt_set_refresh_token_manually}
\alias{qt_set_refresh_token_manually}
\title{Manually set the refresh token.}
\usage{
qt_set_refresh_token_manually(account_set = load_account_set(), value = NULL)
}
\arguments{
\item{account_set}{an account_set object}

\item{value}{The value of the refresh token. Leave NULL to prompt the user for this.}
}
\description{
Use this to set the refresh token directly. You can get a refresh token from the Questrade
API hub page in your Questrade account.
}
