% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{request_refresh_token}
\alias{request_refresh_token}
\title{Request refresh token}
\usage{
request_refresh_token(account_set, request_token)
}
\arguments{
\item{account_set}{account set object}

\item{rt}{valid request_token}
}
\value{
an httr response object from questrade.com containing refreshed tokens
}
\description{
Request refresh token
}
