% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_stored_token}
\alias{set_stored_token}
\alias{get_stored_token}
\title{Get and set stored tokens}
\usage{
set_stored_token(
  account_set,
  token_type = c(REFRESH_TOKEN, ACCESS_TOKEN),
  value = NULL
)

get_stored_token(account_set, token_type = c(REFRESH_TOKEN, ACCESS_TOKEN))
}
\arguments{
\item{account_set}{an account_set object}

\item{token_type}{what type of token to store? Either REFRESH_TOKEN or ACCESS_TOKEN.}

\item{value}{The value to store. If blank, the user will be prompted for a token.}
}
\value{
get_stored_token returns an account_set object
}
\description{
Get and set stored tokens

Set stored token
}
\section{Functions}{
\itemize{
\item \code{get_stored_token}: get stored tokens
}}

